requirements.txt
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blinker==1.5
blis==0.7.9
braceexpand==0.1.7
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
decord==0.6.0
distlib==0.3.6
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fairscale==0.4.4
fastapi==0.92.0
filelock==3.9.0
fonttools==4.38.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.12.1
identify==2.5.18
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
iopath==0.1.10
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kaggle==1.5.12
kiwisolver==1.4.4
langcodes==3.3.0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
murmurhash==1.0.9
networkx==3.0
nodeenv==1.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
portalocker==2.7.0
pre-commit==3.0.4
preshed==3.0.8
prompt-toolkit==3.0.37
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
python-multipart==0.0.5
python-slugify==8.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rich==13.3.1
salesforce-lavis==1.0.0
scikit-image==0.19.3
scipy==1.10.1
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
starlette==0.25.0
streamlit==1.18.1
tenacity==8.2.1
text-unidecode==1.3
thinc==8.1.7
tifffile==2023.2.3
timm==0.4.12
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.19.0
wasabi==1.1.1
watchdog==2.2.1
watchfiles==0.18.1
wcwidth==0.2.6
webdataset==0.2.33
websockets==10.4
zipp==3.14.0